requirements.txt
fastapi
uvicorn
python-dotenv
vonage==2.5.2
sqlalchemy
aiosqlite
